Added rating to Meal Card and Meal Detail Page

+++ lib/meal_model.dart
@@ -8,7 +8,7 @@ class Meal {
   final String price;
   String imageUrl;
 
-  int rating = 0;
+  double rating = 0;
 
   Meal(this.name, this.location, this.description, this.price);
